parse_keyword -> command_to_snippet and move to central location
