+++ latex_own_command_completions.py
@@ -1,13 +1,14 @@
 import sublime
-import re
 
 _ST3 = sublime.version() >= '3000'
 
 if _ST3:
     from .latextools_utils import analysis, cache
+    from .latextools_utils.parser_utils import command_to_snippet
     from .getTeXRoot import get_tex_root
 else:
     from latextools_utils import analysis, cache
+    from latextools_utils.parser_utils import command_to_snippet
     from getTeXRoot import get_tex_root
 
 __all__ = ["get_own_env_completion", "get_own_command_completion"]
@@ -79,31 +80,8 @@ def _parse_command(c):
         elif arg_count == 0:
             raise NoArgs()
         s += "{arg}" * arg_count
-        comp = parse_keyword(s)
+        comp = command_to_snippet(s)
     except:  # no args
         s = c.args + "{}"
         comp = s
-    return (s + "\tself-defined", comp)
-
-
-# TODO duplicate to avoid cyclic imports
-def parse_keyword(keyword):
-    # Replace strings in [] and {} with snippet syntax
-    def replace_braces(matchobj):
-        replace_braces.index += 1
-        if matchobj.group(1) is not None:
-            word = matchobj.group(1)
-            return u'{${%d:%s}}' % (replace_braces.index, word)
-        else:
-            word = matchobj.group(2)
-            return u'[${%d:%s}]' % (replace_braces.index, word)
-    replace_braces.index = 0
-
-    replace, n = re.subn(r'\{([^\{\}\[\]]*)\}|\[([^\{\}\[\]]*)\]', replace_braces, keyword)
-
-    # I do not understand why some of the input will eat the '\' charactor before it!
-    # This code is to avoid these things.
-    if n == 0 and re.search(r'^[a-zA-Z]+$', keyword[1:].strip()) != None:
-        return keyword
-    else:
-        return replace
+    return (s + "\tlocal", comp)
